Change jason->yaml, upgrade Choose _command.kt, add 5 commands

+++ src/main/kotlin/Coordinates.kt
@@ -1,11 +1,11 @@
 import kotlinx.serialization.Serializable
 
 @Serializable
-class Coordinates (private val x: Int, private val y:Int)  {
-    fun get_x(): Int{
+class Coordinates (private val x: Long, private val y:Long)  {
+    fun get_x(): Long{
         return x
     }
-    fun get_y(): Int{
+    fun get_y(): Long{
         return y
     }
 }
